+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Project Changelog
 =================
 
+Release 1.1.0 (30 July 2020)
+----------------------------
+
+* Fix mesh loading when SOLPSPlasma is unpickled.
+* Support Raysect 0.6 and Cherab-core 1.2.
+* Add support for multiple versions of Eirene fort.44 files.
+* Add support for loading a simulation from a balance.nc file.
+* Fix MDSplus neutral radiation and density loading.
+* Add boron to the species list when creating a plasma from a SOLPSSimulation.
+
 Release 1.0.0 (13 Sept 2017)
 ----------------------------